requirements.txt
Django==2.1.4
django-appconf==1.0.2
django-imagekit==4.0.2
model-mommy==1.6.0
pilkit==2.0
Pillow==5.3.0
pytz==2018.7
six==1.11.0
